chore: Remove maven docker image from build
The use of the maven-3.3.3 Docker image in the build created
a memory leak and was causing the CPU to spike to 100% for long
periods of time.  After a few successful builds, the builds
would fail with OutOfMemoryExceptions.

Removing the use of Docker in Docker containers so maven no longer
runs as a container inside of the jenkins:blueocean Docker
container to complete the build.

The Jenkins pipeline will require mvn installed on the system.